adsphere-ai/requirements.txt
fastapi~=0.115.12
uvicorn[standard]~=0.34.2
pika~=1.3.2
python-dotenv~=1.1.0
pymongo~=4.13.0
requests~=2.32.3
python-slugify~=8.0.4
python-dateutil~=2.9.0
pydantic~=2.11.5
transformers~=4.52.4
torch~=2.7.1
pandas~=2.3.0
scikit-learn~=1.6.1
matplotlib~=3.10.3
joblib~=1.5.1
openai~=1.93.0
imagehash~=4.3.1